# repsy-backend/src/main/resources/openapi/openapi-spec.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
paths:
...
  "/api/security/scans":
    get:
      tags:
        - security-scan-controller
      operationId: listSecurityScans
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: severity
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/Severity"
        - name: repoType
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/RepoType"
        - name: repoName
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RestResponsePagedModelVulnerabilityScanInfo"
        '401':
          description: Missing or invalid token, or the authenticated user is not an admin
      summary: List vulnerability scans across all repositories, filterable by severity/repoType/repoName
  "/api/security/scans/summary":
    get:
      tags:
        - security-scan-controller
      operationId: getSecurityScansSummary
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: repoType
          in: query
          required: false
          schema:
            "$ref": "#/components/schemas/RepoType"
        - name: repoName
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RestResponseSecurityScansSummary"
        '401':
          description: Missing or invalid token, or the authenticated user is not an admin
      summary: Get the system-wide severity distribution (each artifact-version's latest scan), filterable by repoType/repoName, independent of the scans list's severity filter
...
components:
  schemas:
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    ResponseType:
      type: string
      enum:
        - ERROR
        - SUCCESS
        - WARNING
...
    RepoType:
      type: string
      enum:
        - MAVEN
        - NPM
        - PYPI
        - DOCKER
        - CARGO
        - NUGET
        - GOLANG
        - HELM
        - RUBY
...
    ScanStatus:
      type: string
      enum:
        - PENDING
        - RUNNING
        - COMPLETED
        - FAILED
    Severity:
      type: string
      enum:
        - CRITICAL
        - HIGH
        - MEDIUM
        - LOW
        - UNKNOWN
...
    VulnerabilityScanInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repoName:
          type: string
        repoType:
          "$ref": "#/components/schemas/RepoType"
        artifactName:
          type: string
        artifactVersion:
          type: string
        status:
          "$ref": "#/components/schemas/ScanStatus"
        highestSeverity:
          "$ref": "#/components/schemas/Severity"
        scannerName:
          type: string
          nullable: true
        scannerVersion:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    PagedModelVulnerabilityScanInfo:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/VulnerabilityScanInfo"
        page:
          "$ref": "#/components/schemas/PageMetadata"
    RestResponsePagedModelVulnerabilityScanInfo:
      type: object
      properties:
        msgId:
          type: string
        type:
          "$ref": "#/components/schemas/ResponseType"
        data:
          "$ref": "#/components/schemas/PagedModelVulnerabilityScanInfo"
        errorCode:
          type: string
        text:
          type: string
...
    SecurityScansSummary:
      type: object
      properties:
        criticalCount:
          type: integer
          format: int32
        highCount:
          type: integer
          format: int32
        mediumCount:
          type: integer
          format: int32
        lowCount:
          type: integer
          format: int32
        unknownCount:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
    RestResponseSecurityScansSummary:
      type: object
      properties:
        msgId:
          type: string
        type:
          "$ref": "#/components/schemas/ResponseType"
        data:
          "$ref": "#/components/schemas/SecurityScansSummary"
        errorCode:
          type: string
        text:
          type: string
...